Separated regressions tests
